VaishaliRestaurantApp/backend-py/requirements.txt
Django
djangorestframework
channels
celery
redis
pandas
